scripts/requirements.txt
# Python dependencies for MicDrop scripts
openai>=1.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
